# docs/apis/parrotfy_openapi.yaml
...
openapi: 3.0.1
info:
  title: API V1
  version: v1
paths:
...
  "/api/v1/products":
    get:
      summary: Obtiene una lista de productos
      tags:
      - Productos
      security:
      - api_key: []
      parameters:
      - name: limit
        in: query
        required: false
        description: Número máximo de registros a retornar (por defecto 20)
        schema:
          type: integer
      - name: active
        in: query
        required: false
        description: Filtrar por estado activo (por defecto true)
        schema:
          type: boolean
      responses:
        '200':
          description: retorna un número limitado de registros
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      nullable: false
                      description: Identificador único del producto
                    code:
                      type: string
                      nullable: true
                      description: Código del producto o SKU
                    name:
                      type: string
                      nullable: false
                      description: Nombre del producto
                    notes:
                      type: string
                      nullable: true
                      description: Notas adicionales sobre el producto
                    created_at:
                      type: string
                      format: date-time
                      nullable: false
                      description: Fecha y hora cuando se creó el producto
                    updated_at:
                      type: string
                      format: date-time
                      nullable: false
                      description: Fecha y hora de la última actualización
                    tax_percentage:
                      type: number
                      nullable: false
                      description: Porcentaje de impuesto aplicado al producto
                    finished_cost:
                      type: number
                      nullable: true
                      description: Costo terminado del producto
                    tax_category:
                      type: integer
                      nullable: true
                      description: 'Categoría de impuesto (0: producto, 1: servicio,
                        2: inmune)'
                    active:
                      type: boolean
                      nullable: false
                      description: Indica si el producto está activo
                    ncm:
                      type: string
                      nullable: true
                      description: Código NCM (Nomenclatura Comum do Mercosul) para
                        clasificación fiscal brasileña
                    external_id:
                      type: string
                      nullable: true
                      description: Identificador de sistema externo
                    tax_free:
                      type: boolean
                      nullable: false
                      description: Indica si el producto está libre de impuestos
                    sec_currency_finished_cost:
                      type: number
                      nullable: true
                      description: Costo terminado en moneda secundaria
                    has_sec_currency:
                      type: boolean
                      nullable: false
                      description: Indica si el producto usa moneda secundaria
                    manage_stock:
                      type: boolean
                      nullable: false
                      description: Indica si se gestiona stock para este producto
                    current_stock:
                      type: number
                      nullable: false
                      description: Cantidad de stock actual
                    unit:
                      type: string
                      nullable: true
                      description: Unidad de medida
                    estimated_cost:
                      type: number
                      nullable: true
                      description: Costo estimado
                    sec_currency_estimated_cost:
                      type: number
                      nullable: true
                      description: Costo estimado en moneda secundaria
                    image_file_name:
                      type: string
                      nullable: true
                      description: Nombre del archivo de imagen
                    image_content_type:
                      type: string
                      nullable: true
                      description: Tipo de contenido de la imagen
                    image_file_size:
                      type: integer
                      nullable: true
                      description: Tamaño del archivo de imagen en bytes
                    image_updated_at:
                      type: string
                      format: date-time
                      nullable: true
                      description: Fecha y hora cuando se actualizó la imagen por
                        última vez
                    include_tax:
                      type: boolean
                      nullable: false
                      description: Indica si el impuesto está incluido en el precio
                    price_with_tax:
                      type: number
                      nullable: true
                      description: Precio incluyendo impuesto
                    sec_currency_price_with_tax:
                      type: number
                      nullable: true
                      description: Precio con impuesto en moneda secundaria
                    provider_id:
                      type: integer
                      nullable: true
                      description: Identificador del proveedor asociado
                    min_stock_alert:
                      type: number
                      nullable: false
                      description: Nivel mínimo de stock para alertas
                    restock_suggestion:
                      type: number
                      nullable: false
                      description: Cantidad sugerida para reabastecimiento
                    specific_tax_id:
                      type: integer
                      nullable: true
                      description: Identificador del impuesto específico aplicado
                        al producto
                    available_for_sale:
                      type: boolean
                      nullable: false
                      description: Indica si el producto está disponible para venta
                    external_name:
                      type: string
                      nullable: true
                      description: Nombre del producto en sistema externo
                    purchasing_unit_name:
                      type: string
                      nullable: true
                      description: Nombre de la unidad de compra
                    purchasing_unit_conversion:
                      type: number
                      nullable: false
                      description: Factor de conversión para unidades de compra
                    barcode:
                      type: string
                      nullable: true
                      description: Código de barras del producto
                    height:
                      type: number
                      nullable: false
                      description: Altura del producto
                    width:
                      type: number
                      nullable: false
                      description: Ancho del producto
                    number_of_pieces:
                      type: number
                      nullable: false
                      description: Número de piezas por unidad
                    priced_by_area:
                      type: boolean
                      nullable: false
                      description: Indica si el producto se cobra por área
                    billing_info:
                      type: object
                      nullable: false
                      description: Información de facturación adicional (JSON)
                    pricing_method:
                      type: integer
                      nullable: false
                      description: 'Método de precio (0: por_unidad, 1: por_área,
                        2: por_planchas, 3: por_paquetes)'
                    plank_area:
                      type: number
                      nullable: false
                      description: Área de planchas
                    unit_of_measurement:
                      type: integer
                      nullable: false
                      description: 'Unidad de medida (0: unidad_1, 1: unidad_2)'
                    estimated_purchase_cost:
                      type: number
                      nullable: true
                      description: Costo de compra estimado
                    sec_currency_estimated_purchase_cost:
                      type: number
                      nullable: true
                      description: Costo de compra estimado en moneda secundaria
                    with_minimum_charge_recipe:
                      type: boolean
                      nullable: false
                      description: Indica si el producto tiene receta de cargo mínimo
        '401':
          description: no autorizado
  "/api/v1/products/{id}":
    get:
      summary: Obtiene un producto específico
      tags:
      - Productos
      security:
      - api_key: []
      parameters:
      - name: id
        in: path
        required: true
        description: ID del producto
        schema:
          type: integer
      responses:
        '200':
          description: producto encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    nullable: false
                    description: Identificador único del producto
                  code:
                    type: string
                    nullable: true
                    description: Código del producto o SKU
                  name:
                    type: string
                    nullable: false
                    description: Nombre del producto
                  notes:
                    type: string
                    nullable: true
                    description: Notas adicionales sobre el producto
                  created_at:
                    type: string
                    format: date-time
                    nullable: false
                    description: Fecha y hora cuando se creó el producto
                  updated_at:
                    type: string
                    format: date-time
                    nullable: false
                    description: Fecha y hora de la última actualización
                  tax_percentage:
                    type: number
                    nullable: false
                    description: Porcentaje de impuesto aplicado al producto
                  finished_cost:
                    type: number
                    nullable: true
                    description: Costo terminado del producto
                  tax_category:
                    type: integer
                    nullable: true
                    description: 'Categoría de impuesto (0: producto, 1: servicio,
                      2: inmune)'
                  active:
                    type: boolean
                    nullable: false
                    description: Indica si el producto está activo
                  ncm:
                    type: string
                    nullable: true
                    description: Código NCM (Nomenclatura Comum do Mercosul) para
                      clasificación fiscal brasileña
                  external_id:
                    type: string
                    nullable: true
                    description: Identificador de sistema externo
                  tax_free:
                    type: boolean
                    nullable: false
                    description: Indica si el producto está libre de impuestos
                  sec_currency_finished_cost:
                    type: number
                    nullable: true
                    description: Costo terminado en moneda secundaria
                  has_sec_currency:
                    type: boolean
                    nullable: false
                    description: Indica si el producto usa moneda secundaria
                  manage_stock:
                    type: boolean
                    nullable: false
                    description: Indica si se gestiona stock para este producto
                  current_stock:
                    type: number
                    nullable: false
                    description: Cantidad de stock actual
                  unit:
                    type: string
                    nullable: true
                    description: Unidad de medida
                  estimated_cost:
                    type: number
                    nullable: true
                    description: Costo estimado
                  sec_currency_estimated_cost:
                    type: number
                    nullable: true
                    description: Costo estimado en moneda secundaria
                  image_file_name:
                    type: string
                    nullable: true
                    description: Nombre del archivo de imagen
                  image_content_type:
                    type: string
                    nullable: true
                    description: Tipo de contenido de la imagen
                  image_file_size:
                    type: integer
                    nullable: true
                    description: Tamaño del archivo de imagen en bytes
                  image_updated_at:
                    type: string
                    format: date-time
                    nullable: true
                    description: Fecha y hora cuando se actualizó la imagen por última
                      vez
                  include_tax:
                    type: boolean
                    nullable: false
                    description: Indica si el impuesto está incluido en el precio
                  price_with_tax:
                    type: number
                    nullable: true
                    description: Precio incluyendo impuesto
                  sec_currency_price_with_tax:
                    type: number
                    nullable: true
                    description: Precio con impuesto en moneda secundaria
                  provider_id:
                    type: integer
                    nullable: true
                    description: Identificador del proveedor asociado
                  min_stock_alert:
                    type: number
                    nullable: false
                    description: Nivel mínimo de stock para alertas
                  restock_suggestion:
                    type: number
                    nullable: false
                    description: Cantidad sugerida para reabastecimiento
                  specific_tax_id:
                    type: integer
                    nullable: true
                    description: Identificador del impuesto específico aplicado al
                      producto
                  available_for_sale:
                    type: boolean
                    nullable: false
                    description: Indica si el producto está disponible para venta
                  external_name:
                    type: string
                    nullable: true
                    description: Nombre del producto en sistema externo
                  purchasing_unit_name:
                    type: string
                    nullable: true
                    description: Nombre de la unidad de compra
                  purchasing_unit_conversion:
                    type: number
                    nullable: false
                    description: Factor de conversión para unidades de compra
                  barcode:
                    type: string
                    nullable: true
                    description: Código de barras del producto
                  height:
                    type: number
                    nullable: false
                    description: Altura del producto
                  width:
                    type: number
                    nullable: false
                    description: Ancho del producto
                  number_of_pieces:
                    type: number
                    nullable: false
                    description: Número de piezas por unidad
                  priced_by_area:
                    type: boolean
                    nullable: false
                    description: Indica si el producto se cobra por área
                  billing_info:
                    type: object
                    nullable: false
                    description: Información de facturación adicional (JSON)
                  pricing_method:
                    type: integer
                    nullable: false
                    description: 'Método de precio (0: por_unidad, 1: por_área, 2:
                      por_planchas, 3: por_paquetes)'
                  plank_area:
                    type: number
                    nullable: false
                    description: Área de planchas
                  unit_of_measurement:
                    type: integer
                    nullable: false
                    description: 'Unidad de medida (0: unidad_1, 1: unidad_2)'
                  estimated_purchase_cost:
                    type: number
                    nullable: true
                    description: Costo de compra estimado
                  sec_currency_estimated_purchase_cost:
                    type: number
                    nullable: true
                    description: Costo de compra estimado en moneda secundaria
                  with_minimum_charge_recipe:
                    type: boolean
                    nullable: false
                    description: Indica si el producto tiene receta de cargo mínimo
        '401':
          description: no autorizado
        '404':
          description: producto no encontrado
...
servers:
- url: https://telconsulting.parrotfy.com
components:
  securitySchemes:
    api_key:
      type: apiKey
      name: P-API-KEY
      in: header